requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
pymongo
sqlmodel
psycopg2-binary
pgvector
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sentence-transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch
